ecommerce_app/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-xml>=2.0.0
requests>=2.31.0
requests-oauthlib>=1.3.1
Pillow>=10.0.0
python-decouple>=3.8

